requirements.txt
paho-mqtt==1.6.1  
pymongo==3.11.0 
redis==4.3.4